requirements.txt
absl-py==2.0.0
altair==5.0.1
altgraph==0.17.4
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.7
cvlib==0.2.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.12.2
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.41.1
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
imutils==0.5.4
ipykernel==6.25.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.14.0
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
macholib==1.16.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
numpy==1.25.1
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
progressbar==2.5
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.15.1
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
Pympler==1.0.1
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
scikit-image==0.21.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.23
stack-data==0.6.2
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
thop==0.1.1.post2209072238
tifffile==2023.9.26
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.144
urllib3==2.0.4
validators==0.20.0
wcwidth==0.2.6
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.16.2
